requirements.txt
# streamlit run app.py
# python -m venv venv 
# run virtual env: venv\Scripts\activate.bat     


scikit-learn
joblib

beautifulsoup4
requests
lxml
streamlit
reportlab
pandas
numpy
scikit-learn
joblib
python-dotenv
beautifulsoup4
requests
lxml
pypdf 
sentence-transformers
pinecone
langchain
google-generativeai==0.8.5
PyPDF2
